Upgrade hardhat.config.ts
Common parameters for networks are entered in the createNetworkConfig function, which automatically builds the configuration for each network. This eliminates duplication.

All compilers use the same optimizer setting (for most contracts). This is submitted to the COMPILER_SETTINGS constant.

Account configuration is now stored in one ACCOUNTS_CONFIG object, which is used for all networks.

Flexibility is applied through the use of the INFURA_URL variable, which makes it easy to change the URL for providers without having to copy the same code