Upgrade verifyMainnetDeployment.js
Instead of prescribing the path to artifacts for each contract every time, I created an artifactsPath object that stores the path to all the necessary files. This makes it easy to control and change paths in one place.

To improve readability and reduce repeating code, I have taken repeated actions into separate functions

Renamed some variables to improve code readability

Excess or non-informative comments have also been removed, for example, indications of the location of files that are now processed automatically.

---------------------------------------------------

there were typos in the original contract if you decide to leave it: conracts - contracts,  btyecode - bytecode

"deployMainnet.polygonZkEVMBridgeAddress" and other proxy addresses are checked several times. It is better to check each address specifically in the context of his contract to avoid confusion.

Printout:"polygonZkEVMGlobalImpl" must be "polygonZkEVMGlobalExitRootImpl" so that the variable name matches the content.

and few more